+++ src/test/java/chess/domain/state/BoardInitializerTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Piece;
-import chess.domain.state.BoardInitializer;
+import chess.domain.board.BoardInitializer;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
