+++ src/test/java/chess/domain/state/WhiteWinTest.java
@@ -4,11 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Position;
-import chess.domain.state.BoardInitializer;
-import chess.domain.state.GameState;
-import chess.domain.state.WhiteTurn;
-import chess.domain.state.WhiteWin;
-import chess.domain.state.Winner;
+import chess.domain.board.BoardInitializer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
